refactor(modal): extract ModalHeaderShell shared by Project and Experience modals
SonarCloud was flagging cross-file duplication: ProjectModalHeader and
experience/ModalHeader both implemented the same sticky header pattern
(animated motion.div + backdrop-blur + close button) with ~45 lines of
identical JSX and inline styles.

Extract the shell into src/components/ui/ModalHeaderShell.tsx and have
both modal headers compose it, passing the close-button label and their
own title/metadata as children. Behavior is unchanged for both modals.
