+++ src/pages/experience/ModalHeader.tsx
@@ -1,11 +1,6 @@
-import { motion } from "motion/react";
-import { X, Building2 } from "lucide-react";
-import {
-   TEXT_PRIMARY,
-   TEXT_SECONDARY,
-   CYAN,
-   MONO_FONT,
-} from "@/constants/theme";
+import { Building2 } from "lucide-react";
+import ModalHeaderShell from "@components/ui/ModalHeaderShell";
+import { TEXT_PRIMARY, CYAN, MONO_FONT } from "@/constants/theme";
 import type { ProfessionalExperience } from "@/types";
 
 interface ModalHeaderProps {
@@ -15,21 +10,10 @@ interface ModalHeaderProps {
 }
 
 const ModalHeader = ({ experience, onClose, isMobile }: ModalHeaderProps) => (
-   <motion.div
-      initial={{ opacity: 0, y: -20 }}
-      animate={{ opacity: 1, y: 0 }}
-      transition={{ delay: 0.1, duration: 0.4, ease: [0.16, 1, 0.3, 1] }}
-      style={{
-         position: "sticky",
-         top: 0,
-         zIndex: 10,
-         padding: isMobile ? "16px 16px 12px" : "20px 24px 16px",
-         background: "rgba(12,12,28,0.9)",
-         backdropFilter: "blur(20px)",
-         WebkitBackdropFilter: "blur(20px)",
-         borderBottom: "1px solid rgba(255,255,255,0.06)",
-         borderRadius: "20px 20px 0 0",
-      }}
+   <ModalHeaderShell
+      isMobile={isMobile}
+      onClose={onClose}
+      closeLabel="Close experience details"
    >
       <div
          style={{
@@ -58,25 +42,6 @@ const ModalHeader = ({ experience, onClose, isMobile }: ModalHeaderProps) => (
          >
             {experience.company}
          </h3>
-         <button
-            onClick={onClose}
-            aria-label="Close experience details"
-            style={{
-               padding: 8,
-               background: "rgba(255,255,255,0.06)",
-               border: "1px solid rgba(255,255,255,0.1)",
-               borderRadius: 10,
-               cursor: "pointer",
-               color: TEXT_SECONDARY,
-               display: "flex",
-               alignItems: "center",
-               justifyContent: "center",
-               transition: "background 0.2s",
-               flexShrink: 0,
-            }}
-         >
-            <X size={16} />
-         </button>
       </div>
       <p
          style={{
@@ -90,7 +55,7 @@ const ModalHeader = ({ experience, onClose, isMobile }: ModalHeaderProps) => (
       >
          {experience.title} | {experience.date}
       </p>
-   </motion.div>
+   </ModalHeaderShell>
 );
 
 export default ModalHeader;
